// src/types/schemas-auth.d.ts
export interface components {
    schemas: {
// ...
        DiscoverLoginContextsResponse: {
            selectionToken?: string;
            /** Format: int64 */
            expiresInSeconds?: number;
            contexts?: components["schemas"]["DiscoveredLoginContextResponse"][];
        };
        DiscoveredLoginContextResponse: {
            contextId?: string;
            /** Format: uuid */
            tenantId?: string;
            /** Format: uuid */
            userId?: string;
            /** Format: uuid */
            actorId?: string;
            organizations?: components["schemas"]["UserOrganizationAccessResponse"][];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}